+++ lib/multi_split_view.dart
@@ -4,4 +4,3 @@ export 'src/multi_split_view.dart';
 export 'src/theme_widget.dart';
 export 'src/theme_data.dart';
 export 'src/divider_painter.dart';
-export 'src/controller.dart';